Various attempts to make 'Applicative' usable in a constrained setting.
As was already seen previously, this is not quite easy because it's unclear what morphisms as functor arguments are supposed to mean. Introduction of the 'Curry' class allows defining 'Monoidal', which is (at least in Hask) equivalent to 'Applicative'; but there can't seem to arise a good constrained implementation of '<*>'.